+++ CHANGELOG.md
@@ -22,6 +22,7 @@ for git projects.
 
 * Modified `projectile-ack` to append to `ack-and-a-half-arguments`
   instead of overriding them.
+* [#229] Fix `projectile-find-file-in-directory`'s behavior for project directories
 
 ## 0.10.0 (12/09/2013)